Make global example recognized by cargo
