+++ Cargo.toml
@@ -47,3 +47,4 @@ warp = "0.3.1"
 futures = "0.3.15"
 reqwest = "0.11.3"
 glob = "0.3"
+lazy_static = "1"
