--- examples/global/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "global"
-version = "0.1.0"
-edition = "2018"
-
-[dependencies]
-config = { path = "../../" }
-lazy_static = "1"
